// frontend/src/app/api/models/collection-info.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CollectionInfo {
  distance?: (string | null);
  indexedVectors?: (number | null);
  name: string;
  points?: (number | null);
  segments?: (number | null);
  source?: (string | null);
  status?: (string | null);
  vectorSize?: (number | null);
}